task: moving writer and tests to an internal response package and updating writer calls
