+++ core/server/router_test.go
@@ -1,13 +1,11 @@
 package server_test
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"fmt"
 	"io"
 	"log"
-	"net"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -462,70 +460,4 @@ func TestRegisterErrorHandler(t *testing.T) {
 			t.Errorf("Expected 'This is the custom internal server error page :D : runtime error: index out of range [1] with length 0', got %v", body)
 		}
 	})
-}
-
-// testHijackResponseWriter wraps the httptest.ResponseRecorder to implement Hijack interface.
-type testHijackResponseWriter struct {
-	*httptest.ResponseRecorder
-}
-
-func (t *testHijackResponseWriter) Hijack() (net.Conn, *bufio.ReadWriter, error) {
-	return nil, nil, nil
-}
-
-func TestResponseWriter(t *testing.T) {
-	s := server.New(
-		server.WithSession("test_secret", "test"),
-	)
-
-	s.HandleFunc("GET /flush/{$}", func(w http.ResponseWriter, _ *http.Request) {
-		f, ok := w.(http.Flusher)
-		if !ok {
-			w.Write([]byte("Flush not supported"))
-			return
-		}
-
-		f.Flush()
-
-		w.Write([]byte("Flush supported!"))
-	})
-
-	s.HandleFunc("GET /hijack/{$}", func(w http.ResponseWriter, _ *http.Request) {
-		h, ok := w.(http.Hijacker)
-		if !ok {
-			w.Write([]byte("Hijack not supported"))
-			return
-		}
-
-		_, _, err := h.Hijack()
-		if err != nil {
-			w.Write([]byte(err.Error()))
-			return
-		}
-
-		w.Write([]byte("Hijack supported!"))
-	})
-
-	resp := &testHijackResponseWriter{
-		ResponseRecorder: httptest.NewRecorder(),
-	}
-	req := httptest.NewRequest(http.MethodGet, "/flush/", nil)
-	s.Handler().ServeHTTP(resp, req)
-
-	if resp.Body.String() == "Flush not supported" {
-		t.Errorf("Expected 'Flush supported!', got 'Flush not supported'")
-		return
-	}
-
-	resp = &testHijackResponseWriter{
-		ResponseRecorder: httptest.NewRecorder(),
-	}
-
-	req = httptest.NewRequest(http.MethodGet, "/hijack/", nil)
-	s.Handler().ServeHTTP(resp, req)
-
-	if resp.Body.String() == "Hijack not supported" {
-		t.Errorf("Expected 'Hijack supported!', got 'Hijack not supported'")
-		return
-	}
 }
